+++ core/src/main/kotlin/com/katanox/tabour/consumption/Config.kt
@@ -1,0 +1,3 @@
+package com.katanox.tabour.consumption
+
+internal interface Config
