+++ core/src/main/kotlin/com/katanox/tabour/sqs/config/SqsConsumer.kt
@@ -12,22 +12,6 @@ import java.time.Duration
 import java.time.temporal.ChronoUnit
 import software.amazon.awssdk.services.sqs.model.Message
 
-class SqsPipeline : Config {
-    internal var transformerWasSet = false
-    internal var producerWasSet = false
-
-    var producer: SqsProducer = sqsProducer {}
-        set(value) {
-            producerWasSet = value.urlWasSet()
-            field = value
-        }
-
-    var transformer: (Message) -> String = { "" }
-        set(value) {
-            field = value
-            transformerWasSet = true
-        }
-}
 
 class SqsConsumer internal constructor() : Consumer<Message, ConsumptionError>, Config {
     private var onSuccessWasSet: Boolean = false
@@ -38,7 +22,7 @@ class SqsConsumer internal constructor() : Consumer<Message, ConsumptionError>, 
     override var onSuccess: (Message) -> Unit = {}
         set(value) {
             if (pipelineWasSet) {
-                throw IllegalArgumentException("Can not set onSuccess if pipeline is set")
+                throw IllegalArgumentException("Can not set `onSuccess` if pipeline is set")
             }
             onSuccessWasSet = true
             field = value
@@ -60,57 +44,3 @@ class SqsConsumer internal constructor() : Consumer<Message, ConsumptionError>, 
     var config: SqsConsumerConfiguration = sqsConsumerConfiguration { maxMessages = 1 }
 }
 
-class SqsConsumerConfiguration internal constructor() : Config {
-    /** The number of max messages to fetch. Default is 1 with max being 10 */
-    var maxMessages: Int = 1
-        set(value) {
-            if (value > 10 || value < 0) {
-                throw IllegalArgumentException("Value must be 0-10")
-            }
-            field = value
-        }
-
-    /**
-     * Unit of concurrency. In combination with [maxMessages] determines the max number of messages
-     * that can be received every [sleepTime]. The upper bound of messages every [sleepTime] is
-     * [maxMessages] * [concurrency]
-     */
-    var concurrency: Int = 1
-        set(value) {
-            if (value > 10 || value < 0) {
-                throw IllegalArgumentException("Value must be 0-10")
-            }
-            field = value
-        }
-    /**
-     * The delay between subsequent requests
-     *
-     * Default is 10 seconds
-     */
-    var sleepTime: Duration = Duration.of(10L, ChronoUnit.SECONDS)
-
-    /**
-     * The duration for which the call waits for a message to arrive in the queue before returning.
-     * If a message is available, the call returns sooner than WaitTimeSeconds. If no messages are
-     * available and the wait time expires, the call returns successfully with an empty list of
-     * messages.
-     *
-     * Default is 10 seconds
-     */
-    var waitTime: Duration = Duration.of(0L, ChronoUnit.SECONDS)
-
-    /**
-     * The number of attempts to receive a message if an exception occurs
-     *
-     * Default is 1
-     */
-    var retries: Int = 1
-        set(value) {
-            if (value < 0) {
-                throw IllegalArgumentException("Negative values are not allowed")
-            }
-            field = value
-        }
-
-    var consumeWhile: () -> Boolean = { true }
-}
