+++ core/src/main/kotlin/com/katanox/tabour/sqs/production/SqsProducerConfiguration.kt
@@ -1,0 +1,7 @@
+package com.katanox.tabour.sqs.production
+
+import com.katanox.tabour.consumption.Config
+
+class SqsProducerConfiguration internal constructor() : Config {
+    var retries: Int = 1
+}
