+++ core/src/main/kotlin/com/katanox/tabour/sqs/production/TabourProducer.kt
@@ -1,0 +1,5 @@
+package com.katanox.tabour.sqs.production
+
+interface TabourProducer {
+    val key: String
+}
